+++ src/hooks/integration/useChatSSE.ts
