+++ src/lib/utils/scheduleUtils.ts
